+++ src/PIL/Image.py
@@ -57,7 +57,7 @@ def __getattr__(name):
     if name in categories:
         warnings.warn(
             "Image categories are deprecated and will be removed in Pillow 10 "
-            "(2023-01-02). Use is_animated instead.",
+            "(2023-07-01). Use is_animated instead.",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -513,7 +513,7 @@ class Image:
         if name == "category":
             warnings.warn(
                 "Image categories are deprecated and will be removed in Pillow 10 "
-                "(2023-01-02). Use is_animated instead.",
+                "(2023-07-01). Use is_animated instead.",
                 DeprecationWarning,
                 stacklevel=2,
             )