+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog (Pillow)
 9.0.0 (unreleased)
 ------------------
 
+- Use title for display in ImageShow #5788
+  [radarhere]
+
 - Remove support for FreeType 2.7 and older #5777
   [hugovk, radarhere]
 
@@ -389,7 +392,7 @@ Changelog (Pillow)
 - Changed Image.open formats parameter to be case-insensitive #5250
   [Piolie, radarhere]
 
-- Deprecate Tk/Tcl 8.4, to be removed in Pillow 10 (2023-01-02) #5216
+- Deprecate Tk/Tcl 8.4, to be removed in Pillow 10 (2023-07-01) #5216
   [radarhere]
 
 - Added tk version to pilinfo #5226